+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Added `JsonVariant::as<char*>()` as a synonym for `JsonVariant::as<const char*>()` (issue #257)
+* Added example `JsonHttpClient` (issue #256)
 
 v5.1.1
 ------